Define 'user' object on 'req.session'
Replaced 'userId' property with 'user' object that holds id, username
and email of authenticated user. This change was made for more efficient
retrieval of authenticated user's data, that will be needed for
frontend.

The 'logIn' method of service inside 'sessions' module now returns whole
'user' object instead just ID and controller sends it in the response.